+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Forbid single-task edit if task group has origin different from 'other' (\#1040).
 * Fix user group visibility for standard user on templates creation (\#1041).
 * Run svelte-check on `sandbox/` and `tasks-list/` (\#1042).
+* Complete the `templates-list/` app (\#1037).
 
 # 1.24.0